pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[project]
name = "example_py_multiple_dispatch"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "plum-dispatch==2.1.0",
    "pytest",
]
[tool.setuptools.packages.find]
where = ["src"]
